JavaHttpProvider:  clean up tests for javadocs examples.
